Add cwAlarm descriptions & more actions
When using the alarms for notifications it definitely helps to have a
brief regular language description to get an idea of what's happening.

Additionally, we want to receive notifications for when alarms clear as
well as when they alarm or insufficient data is available.

Finally, given we can set the cadence of the snapshots, we need the
alarms to ignore missing data points to avoid false alarms.