Remove redundant call to fetch tags.
This solves a problem where the fuction errors out after being throttled
for too many ListTagsForResource calls.

Tags are fetched again in the if blocks so there's no need to fetch them
at the top of this function. This also will skip tag fetches for
snapshots that have been created automatically.