Add more flexibility in snapshot naming.
Creating a second copy of a snapshot in the same account and same
region means we need to make sure the names are unique. copy_local()
now takes an option for a new name.

That means parsing the snapshot name for a timestamp needs to be more
flexible. get_timestamp() now looks for a timestamp-like string in the
name and doesn't assume there's nothing after the timestamp.